refactor: integrate fact tracing into the current structure
