+++ quanda/benchmarks/downstream_eval/mrr.py
@@ -1,18 +1,16 @@
 """Mean Reciprocal Rank (MRR) benchmark."""
 
 import logging
-from typing import Any, Callable, List, Optional, Union
 
-import datasets  # type: ignore
-import torch
-
-from quanda.benchmarks.base import Benchmark
+from quanda.benchmarks.downstream_eval._fact_tracing import (
+    FactTracingBenchmark,
+)
 from quanda.metrics.downstream_eval.mrr import MRRMetric
 
 logger = logging.getLogger(__name__)
 
 
-class MRR(Benchmark):
+class MRR(FactTracingBenchmark):
     """Benchmark for Mean Reciprocal Rank (MRR) metric.
 
     This benchmark evaluates whether retrieved examples (proponents) logically
@@ -29,96 +27,12 @@ class MRR(Benchmark):
     """
 
     name: str = "Mean Reciprocal Rank"
-    eval_args: list = ["explanations", "entailment_labels"]
 
-    def __init__(
-        self,
-        *args,
-        **kwargs,
-    ):
-        """Initialize the MRR benchmark.
-
-        This initializer is not used directly, instead,
-        the `generate` or the `assemble` methods should be used.
-        Alternatively, `download` can be used to load a precomputed benchmark.
-        """
-        super().__init__()
-
-        self.model: torch.nn.Module
-        self.device: str
-        self.train_dataset: Union[torch.utils.data.Dataset, datasets.Dataset]
-        self.eval_dataset: Union[torch.utils.data.Dataset, datasets.Dataset]
-        self.checkpoints: List[str]
-        self.checkpoints_load_func: Callable[..., Any]
-        self.entailment_labels: Optional[torch.Tensor]
-
-    @classmethod
-    def from_config(
-        cls,
-        config: dict,
-        load_meta_from_disk: bool = True,
-        offline: bool = False,
-        device: str = "cpu",
-    ):
-        """Initialize the benchmark from a dictionary.
-
-        Parameters
-        ----------
-        config : dict
-            Dictionary containing the configuration.
-        load_meta_from_disk : str
-            Loads dataset metadata from disk if True, otherwise generates it,
-            default True.
-        offline : bool
-            If True, the model is not downloaded, default False.
-        device: str, optional
-            Device to use for the evaluation, by default "cpu".
-
-        """
-        obj = super().from_config(config, load_meta_from_disk, offline, device)
-        obj.entailment_labels = obj.train_dataset.entailment_labels
-        return obj
-
-    def evaluate(
-        self,
-        explainer_cls: type,
-        expl_kwargs: Optional[dict] = None,
-        batch_size: int = 8,
-    ):
-        """Evaluate the benchmark using a given explanation method.
-
-        Parameters
-        ----------
-        explainer_cls: type
-            The explanation class inheriting from the base Explainer class to
-            be used for evaluation.
-        expl_kwargs: Optional[dict], optional
-            Keyword arguments for the explainer, by default None.
-        batch_size: int, optional
-            Batch size for the evaluation, by default 8.
-
-        Returns
-        -------
-        Dict[str, float]
-            Dictionary containing the MRR score.
-
-        """
-        explainer = self._prepare_explainer(
-            dataset=self.train_dataset,
-            explainer_cls=explainer_cls,
-            expl_kwargs=expl_kwargs,
-        )
-
-        metric = MRRMetric(
+    def _build_metric(self) -> MRRMetric:
+        """Instantiate the MRR metric bound to this benchmark's assets."""
+        return MRRMetric(
             model=self.model,
             checkpoints=self.checkpoints,
             train_dataset=self.train_dataset,
             checkpoints_load_func=self.checkpoints_load_func,
-        )
-
-        return self._evaluate_dataset(
-            eval_dataset=self.eval_dataset,
-            explainer=explainer,
-            metric=metric,
-            batch_size=batch_size,
         )
