+++ quanda/benchmarks/downstream_eval/recall_at_k.py
@@ -6,13 +6,15 @@ from typing import Any, Callable, List, Optional, Union
 import datasets  # type: ignore
 import torch
 
-from quanda.benchmarks.base import Benchmark
+from quanda.benchmarks.downstream_eval._fact_tracing import (
+    FactTracingBenchmark,
+)
 from quanda.metrics.downstream_eval.recall_at_k import RecallAtKMetric
 
 logger = logging.getLogger(__name__)
 
 
-class RecallAtK(Benchmark):
+class RecallAtK(FactTracingBenchmark):
     """Benchmark for Recall@k metric.
 
     This benchmark evaluates whether retrieved examples (proponents) logically
@@ -29,99 +31,56 @@ class RecallAtK(Benchmark):
     """
 
     name: str = "Recall@k"
-    eval_args: list = ["explanations", "entailment_labels"]
 
     def __init__(
         self,
-        *args,
-        **kwargs,
+        model: torch.nn.Module,
+        train_dataset: Union[torch.utils.data.Dataset, datasets.Dataset],
+        eval_dataset: torch.utils.data.Dataset,
+        checkpoints: List[str],
+        checkpoints_load_func: Callable[..., Any],
+        device: str = "cpu",
+        val_dataset: Optional[
+            Union[torch.utils.data.Dataset, datasets.Dataset]
+        ] = None,
+        use_predictions: bool = False,
+        entailment_labels: Optional[torch.Tensor] = None,
+        k: int = 10,
     ):
         """Initialize the Recall@k benchmark.
 
-        This initializer is not used directly, instead,
-        the `generate` or the `assemble` methods should be used.
-        Alternatively, `download` can be used to load a precomputed benchmark.
-        """
-        super().__init__()
+        Parameters
+        ----------
+        k : int, optional
+            The k value for Recall@k, by default 10.
 
-        self.model: torch.nn.Module
-        self.device: str
-        self.train_dataset: Union[torch.utils.data.Dataset, datasets.Dataset]
-        self.eval_dataset: Union[torch.utils.data.Dataset, datasets.Dataset]
-        self.checkpoints: List[str]
-        self.checkpoints_load_func: Callable[..., Any]
-        self.k: int
-        self.entailment_labels: Optional[torch.Tensor]
+        All other parameters mirror
+        :class:`~quanda.benchmarks.base.Benchmark`.
+        """
+        super().__init__(
+            model=model,
+            train_dataset=train_dataset,
+            eval_dataset=eval_dataset,
+            checkpoints=checkpoints,
+            checkpoints_load_func=checkpoints_load_func,
+            device=device,
+            val_dataset=val_dataset,
+            use_predictions=use_predictions,
+            entailment_labels=entailment_labels,
+        )
+        self.k: int = k
 
     @classmethod
-    def from_config(
-        cls,
-        config: dict,
-        load_meta_from_disk: bool = True,
-        offline: bool = False,
-        device: str = "cpu",
-    ):
-        """Initialize the benchmark from a dictionary.
+    def _extra_kwargs_from_config(cls, config: dict) -> dict:
+        """Pull ``k`` off the config into ``__init__`` kwargs."""
+        return {"k": config.get("k", 10)}
 
-        Parameters
-        ----------
-        config : dict
-            Dictionary containing the configuration.
-        load_meta_from_disk : str
-            Loads dataset metadata from disk if True, otherwise generates it,
-            default True.
-        offline : bool
-            If True, the model is not downloaded, default False.
-        device: str, optional
-            Device to use for the evaluation, by default "cpu".
-
-        """
-        obj = super().from_config(config, load_meta_from_disk, offline, device)
-        obj.k = config.get("k", 10)
-        obj.entailment_labels = obj.train_dataset.entailment_labels
-        return obj
-
-    def evaluate(
-        self,
-        explainer_cls: type,
-        expl_kwargs: Optional[dict] = None,
-        batch_size: int = 8,
-    ):
-        """Evaluate the benchmark using a given explanation method.
-
-        Parameters
-        ----------
-        explainer_cls: type
-            The explanation class inheriting from the base Explainer class to
-            be used for evaluation.
-        expl_kwargs: Optional[dict], optional
-            Keyword arguments for the explainer, by default None.
-        batch_size: int, optional
-            Batch size for the evaluation, by default 8.
-
-        Returns
-        -------
-        Dict[str, float]
-            Dictionary containing the Recall@k score.
-
-        """
-        explainer = self._prepare_explainer(
-            dataset=self.train_dataset,
-            explainer_cls=explainer_cls,
-            expl_kwargs=expl_kwargs,
-        )
-
-        metric = RecallAtKMetric(
+    def _build_metric(self) -> RecallAtKMetric:
+        """Instantiate the Recall@k metric."""
+        return RecallAtKMetric(
             model=self.model,
             checkpoints=self.checkpoints,
             train_dataset=self.train_dataset,
             checkpoints_load_func=self.checkpoints_load_func,
             k=self.k,
-        )
-
-        return self._evaluate_dataset(
-            eval_dataset=self.eval_dataset,
-            explainer=explainer,
-            metric=metric,
-            batch_size=batch_size,
         )
